requirements.txt
# requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.32.0
apache-airflow==2.8.1
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.3.3
joblib==1.5.2
mysql-connector-python==8.3.0
python-dotenv==1.0.1